Move command line tool from 'cmd' to 'cli'
Changing the naming convention from 'cmd' to 'cli' to better reflect
the purpose and usage of the command line interface component.
All references to 'cmd' have been updated to 'cli' across iggy crate
files, including README.md, source code files, and directory names.
This change aims to provide clarity and consistency within the
project's codebase.
