+++ iggy/src/cli/consumer_group/delete_consumer_group.rs
