+++ iggy/src/cli/consumer_group/get_consumer_group.rs
