+++ iggy/src/cli/partitions/delete_partitions.rs
