+++ iggy/src/cli/personal_access_tokens/create_personal_access_token.rs
@@ -1,6 +1,6 @@
+use crate::cli::utils::personal_access_token_expiry::PersonalAccessTokenExpiry;
 use crate::cli_command::{CliCommand, PRINT_TARGET};
 use crate::client::Client;
-use crate::cmd::utils::personal_access_token_expiry::PersonalAccessTokenExpiry;
 use crate::personal_access_tokens::create_personal_access_token::CreatePersonalAccessToken;
 use anyhow::Context;
 use async_trait::async_trait;
