+++ iggy/src/cli/streams/delete_stream.rs
