+++ iggy/src/cli/topics/create_topic.rs
@@ -1,6 +1,6 @@
+use crate::cli::utils::message_expiry::MessageExpiry;
 use crate::cli_command::{CliCommand, PRINT_TARGET};
 use crate::client::Client;
-use crate::cmd::utils::message_expiry::MessageExpiry;
 use crate::identifier::Identifier;
 use crate::topics::create_topic::CreateTopic;
 use anyhow::Context;
