+++ iggy/src/cli/topics/update_topic.rs
@@ -1,6 +1,6 @@
+use crate::cli::utils::message_expiry::MessageExpiry;
 use crate::cli_command::{CliCommand, PRINT_TARGET};
 use crate::client::Client;
-use crate::cmd::utils::message_expiry::MessageExpiry;
 use crate::identifier::Identifier;
 use crate::topics::update_topic::UpdateTopic;
 use anyhow::Context;
