+++ iggy/src/cli/users/delete_user.rs
