+++ iggy/src/cli_command.rs
@@ -2,7 +2,7 @@ use crate::client::Client;
 use anyhow::{Error, Result};
 use async_trait::async_trait;
 
-pub static PRINT_TARGET: &str = "iggy::cmd::output";
+pub static PRINT_TARGET: &str = "iggy::cli::output";
 
 #[async_trait]
 pub trait CliCommand {
