--- iggy/src/cmd/utils/personal_access_token_expiry.rs
@@ -1,3 +1,0 @@
-use crate::cmd::utils::message_expiry::MessageExpiry;
-
-pub type PersonalAccessTokenExpiry = MessageExpiry;
